Fix first join as a host
